+++ CHANGELOG.md
@@ -4,6 +4,15 @@ Changelog
 next (n/a)
 ----------
 
+0.2 (2026-03-05)
+----------------
+
+- Add option to use a login page
+- Add option to specify a login status page
+- Add option to handle HTMl tables by converting them to JSON
+- Add timeout option
+- Add option to ignore self-signed certificates
+
 0.1 (2022-07-23)
 ----------------